# src/main/java/com/cricket/api/CricketApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.10).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.cricket.api;

import com.cricket.model.TeamDetails;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2019-11-28T15:13:11.741Z")

@Api(value = "cricket", description = "the cricket API")
public interface CricketApi {

    @ApiOperation(value = "Get Team details", nickname = "getTeamDetails", notes = "", response = TeamDetails.class, tags={ "team", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = TeamDetails.class),
        @ApiResponse(code = 400, message = "Invalid ID supplied"),
        @ApiResponse(code = 404, message = "Pet not found"),
        @ApiResponse(code = 405, message = "Validation exception") })
    @RequestMapping(value = "/cricket/{teamId}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<TeamDetails> getTeamDetails(@ApiParam(value = "Team ID", required = true) @PathVariable("teamId") String teamId);

}